+++ docs/RELEASE_NOTES.md
@@ -26,15 +26,30 @@
 
 7. Add Migration guide section which specifies explicitly if there are breaking changes and how to tackle them.
 
-### 3.0.0 - MMM DD, 2020
-[Commits](https://github.com/commercetools/commercetools-sync-java/compare/2.3.0...2.4.0) |
-[Javadoc](https://commercetools.github.io/commercetools-sync-java/v/2.4.0/) |
-[Jar](https://bintray.com/commercetools/maven/commercetools-sync-java/2.4.0)
+-->
+
+### 3.0.0 - Nov 18, 2020
+[Commits](https://github.com/commercetools/commercetools-sync-java/compare/2.3.0...3.0.0) |
+[Javadoc](https://commercetools.github.io/commercetools-sync-java/v/3.0.0/) |
+[Jar](https://bintray.com/commercetools/maven/commercetools-sync-java/3.0.0)
 
 - 🚧 **Breaking Changes** (1)
     - **Product Sync**: `ProductDraft.getState()` is changed from `Reference<State>` to `ResourceIdentifier<State>`, so as a library user you don't need to provide a key field in the id field of the Reference. (Now API and JVM SDK support `ResourceIdentifiers` and it supports id or key as a field). [#589](https://github.com/commercetools/commercetools-sync-java/pull/589)
 
--->
+- 🐞 **Bug Fixes** (1)
+    - **Commons** - Fixed a bug in the duration calculation of decorated retry sphere client `RetrySphereClientDecorator` created by `ClientConfigurationUtils`. [#610](https://github.com/commercetools/commercetools-sync-java/issues/610)
+
+- 🎉 **New Features** (4)
+    - **ShoppingList Sync** - Added support for syncing shopping lists between ctp projects. [#594](https://github.com/commercetools/commercetools-sync-java/issues/594)
+    - **ShoppingList Sync** - Introduced `ShoppingListSyncUtils` which calculates all needed update actions after comparing a `ShoppingList` and a `ShoppingListDraft`. [#594](https://github.com/commercetools/commercetools-sync-java/issues/594)
+    - **ShoppingList Sync** - Introduced `ShoppingListUpdateActionUtils` which contains utils for calculating necessary update actions after comparing individual fields of a `ShoppingList` and a `ShoppingListDraft`. [#594](https://github.com/commercetools/commercetools-sync-java/issues/594)
+    - **ShoppingList Sync** - Introduced `ShoppingListReferenceResolutionUtils` which resolves Type references from a ShoppingList to a ShoppingListDraft. [#594](https://github.com/commercetools/commercetools-sync-java/issues/594)
+
+- 🛠️ **Dependency Updates** (1)
+    - `commercetools-jvm-sdk` `1.53.0` -> [`1.54.0`](http://commercetools.github.io/commercetools-jvm-sdk/apidocs/io/sphere/sdk/meta/ReleaseNotes.html#v1_54_0)
+    - `mockito-junit-jupiter` `3.5.13` ->  [`3.6.0`](https://github.com/mockito/mockito/releases/tag/v3.6.0) 
+    - `org.assertj.assertj-core` `3.17.2` ->  [`3.18.1`](https://assertj.github.io/doc/#assertj-core-3-18-1-release-notes)
+   
 ### 2.3.0 - Oct 15, 2020
 [Commits](https://github.com/commercetools/commercetools-sync-java/compare/2.2.1...2.3.0) |
 [Javadoc](https://commercetools.github.io/commercetools-sync-java/v/2.3.0/) |